+++ R/load.R
@@ -1057,11 +1057,11 @@ loadBio <- function(input, tz){
   if (expect_integer) {
     if (any(colnames(bio) == "Code.space")) {
       aux <- paste(bio$Code.space, "-", bio$Signal)
-      if (any(link <- table(aux) > 1))
-        stopAndReport(ifelse(sum(link) > 1, "Tags ", "Tag "), paste(aux[link], collapse = ", "), ifelse(sum(link) > 1," are ", " is "), "duplicated in the biometrics.")
+      check <- table(aux) > 1
+      prefix <- "Tag"
     } else {
-      if (any(link <- table(bio$Signal) > 1))
-        stopAndReport(ifelse(sum(link) > 1, "Signals ", "Signal "), paste(names(table(bio$Signal))[link], collapse = ", "), ifelse(sum(link) > 1," are ", " is "), "duplicated in the biometrics.")
+      check <- table(bio$Signal) > 1
+      prefix <- "Signal"
     }
   } 
   else {
@@ -1069,15 +1069,17 @@ loadBio <- function(input, tz){
       aux <- unlist(apply(bio, 1, function(x) {
                             paste0(x['Code.space'], '-', splitSignals(x['Signal']))
                            }))
-      if (any(link <- table(aux) > 1))
-        stopAndReport(ifelse(sum(link) > 1, "Tags ", "Tag "), paste(aux[link], collapse = ", "), ifelse(sum(link) > 1," are ", " is "), "duplicated in the biometrics.")
+      check <- table(aux) > 1
+      prefix <- "Tag"
     }
     else {
       aux <- unlist(sapply(bio$Signal, splitSignals))
-      if (any(link <- table(aux) > 1))
-        stopAndReport(ifelse(sum(link) > 1, "Signals ", "Signal "), paste(aux[link], collapse = ", "), ifelse(sum(link) > 1," are ", " is "), "duplicated in the biometrics.")
+      check <- table(aux) > 1
+      prefix <- "Signal"
     }
   }
+  if (any(check))
+    stopAndReport(prefix, ifelse(sum(check) > 1, "s ", " "), paste(names(check)[check], collapse = ", "), ifelse(sum(check) > 1," are ", " is "), "duplicated in the biometrics.")
 
   # check sensor names
   if (!expect_integer) {